requirements.txt
EasyDict==1.7
opencv-python==3.4.1.15
shapely==1.6.4
Cython
scipy
pandas
pyyaml
json_tricks
scikit-image
yacs>=0.1.5
tensorboardX>=1.6
torch>=1.1.0
Pillow>=8.0.1
torchvision>=0.4.0
numpy==1.19.3